all files update.

+++ bin/finder-configs.sh
@@ -1,27 +1,31 @@
 #!/bin/bash -e
 
-localizes=(
-  Applications
-  Documents
-  Downloads
-  Library
-  Movies
-  Music
-  Pictures
-  Public
-  Desktop
-)
+OS_NAME="$(uname | awk '{print tolower($0)}')"
+OS_FULL="$(uname -a)"
+OS_TYPE=
 
-echo "Running delete .localized end .DS_Store ..."
-
-### Home directory.
-for localize in "${localizes[@]}"; do
-  tar_dir=~/${localize}
-  if [ -n "$(ls $tar_dir/.localized 2> /dev/null)" ]; then
-    sudo rm -r $tar_dir/.localized
-    echo ".localized in $tar_dir/ is deleted."
+if [ "${OS_NAME}" == "linux" ]; then
+  if [ $(echo "${OS_FULL}" | grep -c "amzn1") -gt 0 ]; then
+    OS_TYPE="yum"
+  elif [ $(echo "${OS_FULL}" | grep -c "amzn2") -gt 0 ]; then
+    OS_TYPE="yum"
+  elif [ $(echo "${OS_FULL}" | grep -c "el6") -gt 0 ]; then
+    OS_TYPE="yum"
+  elif [ $(echo "${OS_FULL}" | grep -c "el7") -gt 0 ]; then
+    OS_TYPE="yum"
+  elif [ $(echo "${OS_FULL}" | grep -c "Ubuntu") -gt 0 ]; then
+    OS_TYPE="apt"
+  elif [ $(echo "${OS_FULL}" | grep -c "coreos") -gt 0 ]; then
+    OS_TYPE="apt"
+  elif [ $(echo "${OS_FULL}" | grep -c "microsoft-standard-WSL") -gt 0 ]; then
+    # WSL
+    if [ $(cat /etc/os-release | grep -c "Ubuntu") -gt 0 ]; then
+      OS_TYPE="apt"
+    fi
   fi
-done
+elif [ "${OS_NAME}" == "darwin" ]; then
+  OS_TYPE="brew"
+fi
 
 ### Slash directory.
 function SlashDirectory() {
@@ -30,13 +34,43 @@ function SlashDirectory() {
     echo ".localized in /$1$2 is deleted."
   fi
 }
-SlashDirectory "Applications/"
-SlashDirectory "Library/"
-SlashDirectory "Applications/" "Utilities/"
 
-defaults write com.apple.finder AppleShowAllFiles TRUE # Show hidden files.
-defaults write com.apple.desktopservices DSDontWriteNetworkStores True # Do not generate .DS_Store.
-killall Finder
+if [ "${OS_TYPE}" == "brew" ]; then
 
-### delete all .DS_Store
-sudo find / -name ".DS_Store" -depth -exec rm {} \;
+  localizes=(
+    Applications
+    Documents
+    Downloads
+    Library
+    Movies
+    Music
+    Pictures
+    Public
+    Desktop
+  )
+  
+  echo "Running delete .localized end .DS_Store ..."
+  
+  ### Home directory.
+  for localize in "${localizes[@]}"; do
+    tar_dir=~/${localize}
+    if [ -n "$(ls $tar_dir/.localized 2> /dev/null)" ]; then
+      sudo rm -r $tar_dir/.localized
+      echo ".localized in $tar_dir/ is deleted."
+    fi
+  done
+
+  SlashDirectory "Applications/"
+  SlashDirectory "Library/"
+  SlashDirectory "Applications/" "Utilities/"
+  
+  defaults write com.apple.finder AppleShowAllFiles TRUE # Show hidden files.
+  defaults write com.apple.desktopservices DSDontWriteNetworkStores True # Do not generate .DS_Store.
+  killall Finder
+  
+  ### delete all .DS_Store
+  sudo find / -name ".DS_Store" -depth -exec rm {} \;
+
+else
+  echo "Skip this setup phase. [${OS_NAME}]"
+fi
